+++ packages/terra-aggregator/CHANGELOG.md
@@ -3,6 +3,17 @@ Changelog
 
 Unreleased
 ----------
+
+4.44.0 - (March 17, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
+4.43.0 - (March 10, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 4.42.0 - (March 3, 2020)
 ------------------
 ### Changed