+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -6,7 +6,12 @@ Unreleased
 ### Changed
 * Tests updated to reflect input splitting in `terra-date-picker`
 
-=======
+4.31.0 - (March 10, 2020)
+------------------
+### Changed
+* Changed `onBlur` callback to also return the following metadata: dateValue, timeValue and isAmbiguousHour
+* Changed `onChange` and `onChangeRaw` callbacks to return the following metadata as part of the third parameter: iso, inputValue, dateValue, timeValue, isAmbiguousHour, isCompleteValue, isValidValue
+
 4.30.0 - (March 3, 2020)
 ------------------
 ### Changed