+++ packages/terra-form-validation/CHANGELOG.md
@@ -3,6 +3,17 @@ ChangeLog
 
 Unreleased
 ----------
+
+1.28.0 - (March 17, 2020)
+------------------
+### Changed
+* Update docs to use mdx syntax.
+
+1.27.0 - (March 10, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 1.26.0 - (March 3, 2020)
 ------------------
 ### Changed