+++ packages/terra-slide-panel-manager/CHANGELOG.md
@@ -3,6 +3,17 @@ Changelog
 
 Unreleased
 ----------
+
+5.29.0 - (March 17, 2020)
+------------------
+### Fixed
+* Updated jest snapshot for terra-slide-panel change
+
+5.28.0 - (March 10, 2020)
+------------------
+### Changed
+* Minor dependency version bump
+
 5.27.0 - (March 3, 2020)
 ------------------
 ### Changed