+++ packages/terra-slide-panel/CHANGELOG.md
@@ -4,6 +4,14 @@ Changelog
 Unreleased
 ----------
 
+3.24.0 - (March 17, 2020)
+------------------
+### Changed
+* Update docs to use MDX syntax
+
+### Fixed
+* Added key to main content and panel content to prevent unmounting when changing panel position
+
 3.23.0 - (February 18, 2020)
 ------------------
 ### Added