+++ packages/terra-theme-provider/CHANGELOG.md
@@ -3,6 +3,14 @@ Changelog
 
 Unreleased
 ----------
+
+4.0.0 - (March 17, 2020)
+------------------
+### Breaking Changes
+* Removed `isGlobalTheme` prop
+* Changed default behavior to always apply theme on the 'html' element instead of the DIV element wrapping the children
+* Converted "Theme Provider" into a functional component
+
 3.20.0 - (March 3, 2020)
 ------------------
 ### Changed